+++ android/app/src/main/java/com/mulkkam/data/repository/IntakeRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package com.mulkkam.data.repository
 
-import MulKkamError
 import com.mulkkam.data.remote.model.request.IntakeAmountRequest
 import com.mulkkam.data.remote.model.request.IntakeHistoryRequest
 import com.mulkkam.data.remote.model.response.toDomain
 import com.mulkkam.data.remote.service.IntakeService
 import com.mulkkam.domain.IntakeHistorySummaries
+import com.mulkkam.domain.MulKkamError
 import com.mulkkam.domain.MulKkamResult
 import com.mulkkam.domain.repository.IntakeRepository
 import java.time.LocalDate
