+++ android/app/src/main/java/com/mulkkam/di/NetworkInjection.kt
@@ -2,7 +2,7 @@ package com.mulkkam.di
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.mulkkam.BuildConfig
-import com.mulkkam.data.remote.MulKkamCallAdapterFactory
+import com.mulkkam.data.remote.adapter.MulKkamCallAdapterFactory
 import com.mulkkam.data.remote.interceptor.AuthorizationInterceptor
 import com.mulkkam.data.remote.service.AuthService
 import com.mulkkam.data.remote.service.CupsService
