+++ android/app/src/main/java/com/mulkkam/domain/MulKkamError.kt
@@ -1,3 +1,5 @@
+package com.mulkkam.domain
+
 sealed class MulKkamError(
     val code: String,
 ) : Throwable() {
