+++ android/app/src/main/java/com/mulkkam/domain/MulKkamResult.kt
@@ -1,7 +1,5 @@
 package com.mulkkam.domain
 
-import MulKkamError
-
 data class MulKkamResult<T>(
     val error: MulKkamError? = null,
     val data: T? = null,
